+++ Classes/Model/MatchReportModel.php
@@ -3,6 +3,7 @@
 namespace System25\T3sports\Model;
 
 use Sys25\RnBase\Configuration\ConfigurationInterface;
+use Sys25\RnBase\Domain\Model\DataInterface;
 use Sys25\RnBase\Utility\Strings;
 use Sys25\RnBase\Utility\T3General;
 use System25\T3sports\Decorator\MatchNoteDecorator;
@@ -285,6 +286,23 @@ class MatchReportModel
     }
 
     /**
+     * Liefert den Namen des Schiedsrichters.
+     */
+    public function getVideoRefereeName($confId = 'matchreport.videoreferee.')
+    {
+        // der Schiedsrichter wird schon als Instanz geliefert.
+        return $this->_getNames2($this->match->getVideoReferee(), $confId);
+    }
+
+    /**
+     * Liefert die Namen der Linienrichters.
+     */
+    public function getVideoAssistNames($confId = 'matchreport.videoassists.')
+    {
+        return $this->_getNames2($this->matchProfileProvider->getVideoAssists($this->match), $confId);
+    }
+
+    /**
      * Liefert den Namen des Heimtrainers.
      */
     public function getCoachNameHome($confId = 'matchreport.coach.')
@@ -514,7 +532,7 @@ class MatchReportModel
     protected function _wrapProfiles($profiles, $confId)
     {
         $ret = [];
-        if (!is_array($profiles)) {
+        if ($profiles instanceof DataInterface || !is_iterable($profiles)) {
             if (!is_object($profiles)) {
                 return [];
             }