+++ Classes/Table/Volleyball/Comparator.php
@@ -7,7 +7,11 @@ use System25\T3sports\Table\IComparator;
 /***************************************************************
  *  Copyright notice
  *
+<<<<<<< HEAD
  *  (c) 2011-2024 Rene Nitzsche (rene@system25.de)
+=======
+ *  (c) 2011-2023 Rene Nitzsche (rene@system25.de)
+>>>>>>> master
  *  All rights reserved
  *
  *  This script is part of the TYPO3 project. The TYPO3 project is